requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
anyio==4.12.1
bitsandbytes==0.49.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
et-xmlfile==2.0.0
filelock==3.24.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
idna==3.11
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pdf2image==1.17.0
pillow==12.1.1
psutil==7.2.2
pygments==2.19.2
python-dateutil==2.9.0.post0
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers @ git+https://github.com/huggingface/transformers.git@0dda1ffbf39c8dabeb4511af86745fb74b10e5aa
triton==3.1.0
typer==0.23.1
typer-slim==0.23.1
typing-extensions==4.15.0
urllib3==2.6.3
python-docx>=0.8.11
aiogram>=3.0.0
json-repair>=0.7.0
